requirements.txt
Django==2.1.8
django-bootstrap-modal-forms==1.4.1
django-widget-tweaks==1.4.2
selenium==3.14.0
pytz==2018.5
django-tables2
django-currentuser
tablib
django-filters


